requirements.txt
requests>=2.31.0
pydantic>=2.7.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
pytest>=8.2.0
google-genai>=1.0.0
google-cloud-documentai>=2.20.0
